AP-13-searching-mechanism
Added the following functionalities:

Counts the listings based on the searched value by the user (both admin and user side).
Added the searching mechanism in the user side.
Issue (care of Danica and Meekah):
Counting mechanism should not be applicable when the user is just browsing the Explore page and the user hasn't searched for anything yet.